# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "EpisodeCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "concept": {
            "default": "",
            "maxLength": 50000,
            "title": "Concept",
            "type": "string"
          },
          "duration_target": {
            "default": 30,
            "exclusiveMinimum": 0.0,
            "maximum": 600.0,
            "title": "Duration Target",
            "type": "number"
          },
          "episode": {
            "anyOf": [
              {
                "maximum": 999.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode"
          },
          "season": {
            "default": 1,
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Season",
            "type": "integer"
          },
          "title": {
            "default": "Épisode sans titre",
            "maxLength": 180,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "title": "EpisodeCreateRequest",
        "type": "object"
      },
...
      "EpisodePatchRequest": {
        "additionalProperties": false,
        "properties": {
          "characters": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Characters"
          },
          "duration_target": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 600.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Target"
          },
          "locations": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locations"
          },
          "logline": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logline"
          },
          "narrative_source": {
            "anyOf": [
              {
                "maxLength": 50000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative Source"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EpisodeStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "story": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EpisodeStory"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 180,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "EpisodePatchRequest",
        "type": "object"
      },
      "EpisodeStatus": {
        "enum": [
          "idea",
          "writing",
          "review",
          "draft",
          "approved",
          "breakdown",
          "production",
          "final"
        ],
        "title": "EpisodeStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NarrativeGenerateRequest": {
        "additionalProperties": false,
        "properties": {
          "model": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt": {
            "default": "",
            "maxLength": 10000,
            "title": "Prompt",
            "type": "string"
          },
          "source_text": {
            "default": "",
            "maxLength": 50000,
            "title": "Source Text",
            "type": "string"
          }
        },
        "title": "NarrativeGenerateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "La Serre",
    "version": "0.2.13"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/episodes": {
      "get": {
        "operationId": "list_episodes_api_episodes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Episodes Api Episodes Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Episodes",
        "tags": [
          "episodes"
        ]
      },
      "post": {
        "operationId": "create_episode_api_episodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Episode Api Episodes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Episode",
        "tags": [
          "episodes"
        ]
      }
    },
...
    "/api/episodes/{episode_id}": {
      "delete": {
        "operationId": "delete_episode_api_episodes__episode_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Episode Api Episodes  Episode Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Episode",
        "tags": [
          "episodes"
        ]
      },
      "get": {
        "operationId": "get_episode_api_episodes__episode_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Episode Api Episodes  Episode Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Episode",
        "tags": [
          "episodes"
        ]
      },
      "put": {
        "operationId": "update_episode_api_episodes__episode_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Episode Api Episodes  Episode Id  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Episode",
        "tags": [
          "episodes"
        ]
      }
    },
    "/api/episodes/{episode_id}/approve": {
      "post": {
        "operationId": "approve_episode_api_episodes__episode_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Episode Api Episodes  Episode Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Episode",
        "tags": [
          "episodes"
        ]
      }
    },
    "/api/episodes/{episode_id}/breakdown/apply": {
      "post": {
        "operationId": "apply_episode_breakdown_api_episodes__episode_id__breakdown_apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreakdownApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Apply Episode Breakdown Api Episodes  Episode Id  Breakdown Apply Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Episode Breakdown",
        "tags": [
          "episodes"
        ]
      }
    },
    "/api/episodes/{episode_id}/breakdown/generate": {
      "post": {
        "operationId": "generate_episode_breakdown_api_episodes__episode_id__breakdown_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Episode Breakdown Api Episodes  Episode Id  Breakdown Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Episode Breakdown",
        "tags": [
          "episodes"
        ]
      }
    },
    "/api/episodes/{episode_id}/draft/apply": {
      "post": {
        "operationId": "apply_episode_draft_api_episodes__episode_id__draft_apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeDraftApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Apply Episode Draft Api Episodes  Episode Id  Draft Apply Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Episode Draft",
        "tags": [
          "episodes"
        ]
      }
    },
    "/api/episodes/{episode_id}/draft/generate": {
      "post": {
        "operationId": "generate_episode_draft_api_episodes__episode_id__draft_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Episode Draft Api Episodes  Episode Id  Draft Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Episode Draft",
        "tags": [
          "episodes"
        ]
      }
    },
...
    "/api/episodes/{episode_id}/review": {
      "post": {
        "operationId": "review_episode_api_episodes__episode_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Review Episode Api Episodes  Episode Id  Review Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Episode",
        "tags": [
          "episodes"
        ]
      }
    },
...
  }
}